--- src/main/java/capstone/design/control_automation/domain/entity/CameraStatus.java
@@ -1,5 +1,0 @@
-package capstone.design.control_automation.domain.entity;
-
-public enum CameraStatus {
-    NORMAL, BROKEN;
-}
