+++ src/main/java/capstone/design/control_automation/entity/Address.java
@@ -1,4 +1,4 @@
-package capstone.design.control_automation.domain.entity;
+package capstone.design.control_automation.entity;
 
 import static lombok.AccessLevel.PUBLIC;
