+++ src/main/java/capstone/design/control_automation/entity/CameraStatus.java
@@ -1,0 +1,5 @@
+package capstone.design.control_automation.entity;
+
+public enum CameraStatus {
+    NORMAL, BROKEN;
+}
