+++ src/main/java/capstone/design/control_automation/entity/EmergencyStatus.java
@@ -1,4 +1,4 @@
-package capstone.design.control_automation.domain.entity;
+package capstone.design.control_automation.entity;
 
 public enum EmergencyStatus {
     POLICE, FIRE, MEDICAL;
