+++ src/main/java/capstone/design/control_automation/video/controller/VideoSearchRestController.java
@@ -1,5 +1,6 @@
-package capstone.design.control_automation.video;
+package capstone.design.control_automation.video.controller;
 
+import capstone.design.control_automation.video.service.VideoService;
 import capstone.design.control_automation.video.dto.VideoRequest;
 import capstone.design.control_automation.video.dto.VideoResponse;
 import java.util.List;
