+++ src/main/java/capstone/design/control_automation/video/dto/VideoForm.java
@@ -1,8 +1,8 @@
 package capstone.design.control_automation.video.dto;
 
-import capstone.design.control_automation.domain.entity.Camera;
-import capstone.design.control_automation.domain.entity.EmergencyStatus;
-import capstone.design.control_automation.domain.entity.Video;
+import capstone.design.control_automation.entity.Camera;
+import capstone.design.control_automation.entity.EmergencyStatus;
+import capstone.design.control_automation.video.entity.Video;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AccessLevel;
