+++ src/main/java/capstone/design/control_automation/video/repository/VideoRepository.java
@@ -1,8 +1,6 @@
 package capstone.design.control_automation.video.repository;
 
-import capstone.design.control_automation.domain.entity.Video;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import capstone.design.control_automation.video.entity.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
